+++ mne_bids/read.py
@@ -245,14 +245,14 @@ def _handle_electrodes_reading(electrodes_fname, coord_frame, raw, verbose):
     nan_chs = []
     for ch_name, ch_coord in zip(ch_names_raw, ch_locs):
         if any(np.isnan(ch_coord)) and ch_name not in raw.info['bads']:
-            problematic_chs.append(ch_name)
-    if problematic_chs != []:
+            nan_chs.append(ch_name)
+    if nan_chs != []:
         logger.warning("There are problematic channels "
                        "that are not set as 'bad', "
                        "but also the channel location "
                        "is 'n/a' in electrodes.tsv: {}. "
                        "Setting montage without "
-                       "those chs...".format(problematic_chs))
+                       "those chs...".format(nan_chs))
 
     # create mne.DigMontage
     ch_pos = dict(zip(ch_names_raw, np.array(ch_locs)))